Fix mod zips closing when sounds fail to load

+++ arc-core/src/arc/audio/Sound.java
@@ -49,7 +49,7 @@ public class Sound extends AudioSource implements DownloadableAudio{
         try {
             handle = wavLoad(data, data.length);
         } catch (ArcRuntimeException e) {
-            if (e.getMessage().contains("File found")) {
+            if (!(file instanceof ZipFi) && e.getMessage().contains("File found")) {
                 file.delete();
                 throw new ArcRuntimeException(file.absolutePath(), e);
             } else throw e;
